Add files via upload
A python script that simplifies "vcf" files produced by Sniffles2. The simplification if done by removing few columns and extracting the information in the "INFO" column. The end result is a simplified "vcf" file that contains : CHROM, POS, Type of SV, Size of the variation, Total number of read counts, Supporting number of read counts and Allele Frequency.
The script is not optimized but it is working.